Updated get_castable_dtype helper and used it for testing astype (#4501)
